pyproject.toml
[project]
name = "remy"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "fastmcp>=3.2.0",
    "ipykernel>=7.2.0",
    "langchain-mcp-adapters>=0.2.2",
    "langchain-openai>=1.1.12",
    "langgraph>=1.1.4",
    "langgraph-cli[inmem]>=0.4.19",
    "python-dotenv>=1.2.2",
]
